requirements.txt
imageio
matplotlib
numpy
Pillow
tabulate
torch
torchvision
tensorboard
tqdm
gdown
